Update for newest db (D3 v2.2)
Fixed db API changes (was causing the compile errors)
Added optional properties InteractRange and CollisionRadius to CacheEntry.
Updated Witch Doctor Sacrifice to use unit location and cluster location.

+++ fBaseXtensions/Stats/LootStats.cs
@@ -277,7 +277,7 @@ namespace fBaseXtensions.Stats
 			}
 			if (thisPluginItemType == PluginItemTypes.KeyStone)
 			{
-				KeyStoneFragments.Stashed += i.ThisItemStackQuantity;
+				KeyStoneFragments.Stashed += (int)i.ThisItemStackQuantity;
 			}
 
 			switch (i.ACDItem.ItemType)
