+++ fBaseXtensions/XML/MoveItemTag.cs
@@ -377,7 +377,7 @@ namespace fBaseXtensions.XML
 			int iPlayerDynamicID = ZetaDia.Me.CommonData.DynamicId;
 			int iOriginalGameBalanceId = item.ThisBalanceID;
 			int iOriginalDynamicID = item.ThisDynamicID;
-			int iOriginalStackQuantity = item.ThisItemStackQuantity;
+			int iOriginalStackQuantity = (int)item.ThisItemStackQuantity;
 			string sOriginalItemName = item.ThisRealName;
 			string sOriginalInternalName = item.ThisInternalName;
 			PluginItemTypes OriginalPluginItemType = ItemFunc.DetermineItemType(item);
@@ -461,7 +461,7 @@ namespace fBaseXtensions.XML
 					// Check if we combine the stacks, we won't overfill them
 					if ((tempitem.GameBalanceId == iOriginalGameBalanceId) && (tempitem.ItemStackQuantity < tempitem.MaxStackCount))
 					{
-						iLeftoverStackQuantity = (tempitem.ItemStackQuantity + iOriginalStackQuantity) - tempitem.MaxStackCount;
+						iLeftoverStackQuantity = (int)((tempitem.ItemStackQuantity + iOriginalStackQuantity) - tempitem.MaxStackCount);
 						iPointX = tempitem.InventoryColumn;
 						iPointY = tempitem.InventoryRow;
 
@@ -609,7 +609,7 @@ namespace fBaseXtensions.XML
 			int iPlayerDynamicID = ZetaDia.Me.CommonData.DynamicId;
 			int iOriginalGameBalanceId = item.ThisBalanceID;
 			int iOriginalDynamicID = item.ThisDynamicID;
-			int iOriginalStackQuantity = item.ThisItemStackQuantity;
+			int iOriginalStackQuantity = (int)item.ThisItemStackQuantity;
 			string sOriginalItemName = item.ThisRealName;
 			string sOriginalInternalName = item.ThisInternalName;
 			PluginItemTypes OriginalPluginItemType = ItemFunc.DetermineItemType(item);
@@ -690,7 +690,7 @@ namespace fBaseXtensions.XML
 					// Check if we combine the stacks, we won't overfill them
 					if ((tempitem.GameBalanceId == iOriginalGameBalanceId) && (tempitem.ItemStackQuantity < tempitem.MaxStackCount))
 					{
-						iLeftoverStackQuantity = (tempitem.ItemStackQuantity + iOriginalStackQuantity) - tempitem.MaxStackCount;
+						iLeftoverStackQuantity = (int)((tempitem.ItemStackQuantity + iOriginalStackQuantity) - tempitem.MaxStackCount);
 						iPointX = tempitem.InventoryColumn;
 						iPointY = tempitem.InventoryRow;
